requirements.txt
Django==2.1.8
django-bulk-update==2.2.0
django-environ==0.4.5
django-markdownx==2.0.28
Markdown==3.1
mod-wsgi==4.6.5
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2018.9
sqlparse==0.3.0
